Create TensorFlow
TensorFlow works like a computational library for writing new algorithms that involve a large number of tensor operations, since neural networks can be easily expressed as computational graphs they can be implemented using TensorFlow as a series of operations on Tensors. Plus, tensors are N-dimensional matrices which represent your data.
With TensorFlow, we can easily visualize each and every part of the graph which is not an option while using Numpy or SciKit.

We are using TensorFlow daily but indirectly with applications like Google Voice Search or Google Photos. These applications are developed using this library.